requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-community
langchain-ollama
langchain-tavily
chromadb
pypdf
streamlit
python-dotenv